` "User model: user_id maydonini BigIntegerField ga o'tkazish; serializerlar uchun yangi generate_link.py qo'shildi."`

+++ core/apps/accounts/models/user.py
@@ -7,7 +7,7 @@ from ..managers import UserManager
 
 class User(auth_models.AbstractUser):
     phone = models.CharField(max_length=255, unique=True, null=True, blank=True)
-    user_id = models.IntegerField(default=0, unique=True, null=True, blank=True)
+    user_id = models.BigIntegerField(default=0, unique=True, null=True, blank=True)
     username = models.CharField(max_length=255, null=True, blank=True)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
